requirements.txt
django
django_compressor
south

psycopg2
gunicorn
gevent

mega.py
